There are so many errors and issues with dependencies that we may have to make a new frontend
