+++ client/package.json
@@ -12,20 +12,33 @@
     "format": "npx prettier . --write"
   },
   "dependencies": {
+    "axios": "^1.7.7",
+    "boostrap": "^2.0.0",
     "cors": "^2.8.5",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "prop-types": "^15.8.1",
+    "react": "^18.0.11",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^7.0.1",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.81.0",
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@eslint/js": "^9.15.0",
+    "@svgr/plugin-svgo": "^8.1.0",
+    "@svgr/webpack": "^8.1.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.0.1",
     "@types/cors": "^2.8.17",
-    "@types/react": "^18.3.12",
-    "@types/react-dom": "^18.3.1",
+    "@types/node": "^22.10.1",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
     "@typescript-eslint/eslint-plugin": "^8.15.0",
     "@typescript-eslint/parser": "^8.15.0",
     "@vitejs/plugin-react": "^4.3.3",
+    "axios": "^1.7.7",
+    "css-select": "^5.1.0",
     "eslint": "^9.15.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.31.0",
@@ -36,10 +49,16 @@
     "eslint-plugin-react-refresh": "^0.4.14",
     "globals": "^15.12.0",
     "jsdom": "^25.0.1",
+    "nth-check": "^2.1.1",
+    "postcss": "^8.4.49",
     "prettier": "^3.3.3",
+    "react": "^18.0.11",
+    "react-dom": "^18.0.11",
+    "svgo": "^3.3.2",
     "typescript": "~5.6.2",
     "typescript-eslint": "^8.15.0",
     "vite": "^5.4.10",
-    "vitest": "^2.1.5"
+    "vitest": "^2.1.5",
+    "web-vitals": "^2.1.4"
   }
 }
