+++ client/src/App.css
@@ -1,5 +1,5 @@
 /*
-Author(s): Djamel Almabouada, Derek Petersen
+Author(s): Geromy Cunningham, Derek Petersen
 Purpose: Class for Styling the Webpage 
 */
