--- client/src/components/v2website/public/apiurl.env
@@ -1,1 +1,0 @@
-REACT_APP_API_URL=http://yourapi.com
